+++ jest.stub.js
@@ -1,22 +1,19 @@
+global.mockMapGL = {
+    on: jest.fn(),
+    addLayer: jest.fn(),
+    addSource: jest.fn(),
+    addControl: jest.fn(),
+    getLayer: jest.fn(),
+    getSource: jest.fn(),
+    setFeatureState: jest.fn(),
+}
+
+global.mockMap = {
+    getMapGL: () => mockMapGL,
+    setHoverState: jest.fn(),
+}
+
 // https://stackoverflow.com/questions/57943736/how-to-fix-window-url-createobjecturl-is-not-a-function-when-testing-mapbox-gl
 if (typeof window.URL.createObjectURL === 'undefined') {
     window.URL.createObjectURL = () => {}
 }
-
-/*
-// https://github.com/facebook/jest/issues/3449
-class Worker {
-    constructor(stringUrl) {
-        this.url = stringUrl
-        this.onmessage = () => {}
-    }
-
-    postMessage(msg) {
-        this.onmessage(msg)
-    }
-}
-
-if (typeof window.Worker === 'undefined') {
-    window.Worker = Worker
-}
-*/