+++ src/layers/EarthEngine.js
@@ -1,11 +1,22 @@
 import Layer from './Layer'
 import getEarthEngineApi from '../utils/eeapi'
+import {
+    getInfo,
+    getScale,
+    hasClasses,
+    combineReducers,
+    getParamsFromLegend,
+    getHistogramStatistics,
+    getFeatureCollectionProperties,
+} from '../utils/earthengine'
+import { isPoint, featureCollection } from '../utils/geometry'
+import { getBufferGeometry } from '../utils/buffers'
+import { polygonLayer, outlineLayer, pointLayer } from '../utils/layers'
+import { setPrecision } from '../utils/numbers'
 
-const defaultOptions = {
-    url:
-        'https://earthengine.googleapis.com/map/{mapid}/{z}/{x}/{y}?token={token}',
+export const defaultOptions = {
     tokenType: 'Bearer',
-    aggregation: 'none',
+    bandReducer: 'sum',
     popup: '{name}: {value} {unit}',
 }
 
@@ -15,195 +26,242 @@ class EarthEngine extends Layer {
             ...defaultOptions,
             ...options,
         })
-
-        this._legend = options.legend || this.createLegend()
     }
 
-    async getSource() {
+    async addTo(map) {
+        await this.init()
+        await this.createSource()
+        this.createLayers()
+        super.addTo(map)
+        this.onLoad()
+    }
+
+    // EE initialise
+    async init() {
+        this.ee = await getEarthEngineApi()
+        await this.setAuthToken()
+    }
+
+    async createSource() {
         const id = this.getId()
 
-        this._ee = await getEarthEngineApi()
+        this.featureCollection = this.getFeatureCollection()
 
-        await this.setAuthToken()
+        const image = await this.createImage()
 
-        this._eeMap = await this.visualize(this.createImage())
+        const { urlFormat } = await this.visualize(image)
 
-        this.setSource(id, {
+        this.setSource(`${id}-raster`, {
             type: 'raster',
             tileSize: 256,
-            tiles: [this._eeMap.urlFormat],
+            tiles: [urlFormat],
         })
+
+        this.setSource(id, {
+            type: 'geojson',
+            data: featureCollection(this.getFeatures()),
+        })
+    }
+
+    createLayers() {
+        const id = this.getId()
+        const source = id
 
         this.addLayer({
             id: `${id}-raster`,
             type: 'raster',
-            source: id,
+            source: `${id}-raster`,
         })
 
-        return this._source
+        this.addLayer(polygonLayer({ id, source, opacity: 0.9 }), true)
+        this.addLayer(outlineLayer({ id, source }))
+        this.addLayer(
+            pointLayer({
+                id,
+                source: `${id}-points`,
+                radius: 4,
+                color: '#333',
+            })
+        )
     }
 
     // Configures client-side authentication of EE API calls by providing a OAuth2 token to use.
-    setAuthToken = () =>
-        new Promise((resolve, reject) => {
+    setAuthToken() {
+        return new Promise((resolve, reject) => {
+            const { data, initialize } = this.ee
             const { accessToken, tokenType } = this.options
 
             if (accessToken) {
-                accessToken.then(token => {
-                    const { access_token, client_id, expires_in } = token
+                accessToken
+                    .then(token => {
+                        const { access_token, client_id, expires_in } = token
 
-                    this._ee.data.setAuthToken(
-                        client_id,
-                        tokenType,
-                        access_token,
-                        expires_in
-                    )
+                        data.setAuthToken(
+                            client_id,
+                            tokenType,
+                            access_token,
+                            expires_in
+                        )
 
-                    this._ee.data.setAuthTokenRefresher(
-                        this.refreshAccessToken.bind(this)
-                    )
+                        data.setAuthTokenRefresher(this.refreshAccessToken)
 
-                    this._ee.initialize(null, null, resolve)
-                })
+                        initialize(null, null, resolve)
+                    })
+                    .catch(reject)
             }
         })
+    }
 
     // Get OAuth2 token needed to create and load Google Earth Engine layers
-    getAuthToken(callback) {
-        const accessToken = this.options.accessToken
-        if (accessToken) {
-            if (accessToken instanceof Function) {
-                // Callback function returning auth obect
-                accessToken(callback)
-            } else {
-                // Auth token as object
-                callback(accessToken)
+    getAuthToken() {
+        return new Promise((resolve, reject) => {
+            const { accessToken } = this.options
+
+            if (accessToken) {
+                if (accessToken instanceof Function) {
+                    // Callback function returning auth obect
+                    accessToken(resolve)
+                } else {
+                    // Auth token as object
+                    resolve(accessToken)
+                }
             }
-        }
+
+            reject(new Error('No access token in layer options.'))
+        })
     }
 
     // Refresh OAuth2 token when expired
-    refreshAccessToken(authArgs, callback) {
+    refreshAccessToken = async (authArgs, callback) => {
         const { tokenType } = this.options
+        const token = await this.getAuthToken()
 
-        this.getAuthToken(token => {
-            callback({
-                token_type: tokenType,
-                access_token: token.access_token,
-                state: authArgs.scope,
-                expires_in: token.expires_in,
-            })
+        callback({
+            token_type: tokenType,
+            access_token: token.access_token,
+            state: authArgs.scope,
+            expires_in: token.expires_in,
         })
     }
 
-    // Create EE tile layer from params (override for each layer type)
-    createImage() {
-        // eslint-disable-line
-        const options = this.options
+    setFeatures(data = []) {
+        this.setSource(`${this.getId()}-points`, {
+            type: 'geojson',
+            data: featureCollection(data.filter(isPoint)),
+        })
 
-        let eeCollection
-        let eeImage
+        super.setFeatures(data.map(this.createBuffer.bind(this)))
+    }
 
-        if (options.filter) {
-            // Image collection
-            eeCollection = this._ee.ImageCollection(options.datasetId) // eslint-disable-line
+    // Transform point feature to buffer polygon
+    createBuffer(feature) {
+        const { buffer } = this.options
 
-            eeCollection = this.applyFilter(eeCollection)
+        return buffer && feature.geometry.type === 'Point'
+            ? {
+                  ...feature,
+                  geometry: getBufferGeometry(feature, buffer / 1000),
+              }
+            : feature
+    }
 
-            if (options.aggregation === 'mosaic') {
-                this.eeCollection = eeCollection
-                eeImage = eeCollection.mosaic()
-            } else {
-                eeImage = this._ee.Image(eeCollection.first()) // eslint-disable-line
-            }
-        } else {
-            // Single image
-            eeImage = this._ee.Image(options.datasetId) // eslint-disable-line
-        }
+    // Create feature collection for org unit aggregations
+    getFeatureCollection() {
+        const { FeatureCollection } = this.ee
+        const features = this.getFeatures()
 
-        if (options.band) {
-            eeImage = eeImage.select(options.band)
-        }
+        return features.length
+            ? FeatureCollection(
+                  features.map(f => ({
+                      ...f,
+                      id: f.properties.id, // EE requires id to be string, Mapbox integer
+                  }))
+              )
+            : null
+    }
 
-        if (options.mask) {
-            // Mask out 0-values
-            eeImage = eeImage.updateMask(eeImage.gt(0))
-        }
+    // Create EE tile layer from params
+    async createImage() {
+        const { datasetId } = this.options
+
+        // Apply filter (e.g. period)
+        let eeImage = await this.applyFilter(datasetId)
+
+        // Select band (e.g. age group)
+        eeImage = this.selectBand(eeImage)
+
+        // Mask out 0-values
+        eeImage = this.maskImage(eeImage)
 
         // Run methods on image
         eeImage = this.runMethods(eeImage)
 
         this.eeImage = eeImage
 
+        // Image is ready for aggregations
+        this.fire('imageready', { type: 'imageready', image: eeImage })
+
         // Classify image
-        if (!options.legend) {
-            // Don't classify if legend is provided
-            eeImage = this.classifyImage(eeImage)
+        return this.classifyImage(eeImage)
+    }
+
+    // Apply array of filters returns image
+    applyFilter = async datasetId => {
+        const { filter, mosaic } = this.options
+        const { Filter, Image, ImageCollection } = this.ee
+
+        if (!filter) {
+            const image = Image(datasetId)
+            this.scale = await getScale(image)
+            return image
+        }
+
+        let collection = ImageCollection(datasetId)
+
+        // Scale is lost when creating a mosaic below
+        // https://developers.google.com/earth-engine/guides/projections
+        this.scale = await getScale(collection.first())
+
+        filter.forEach(f => {
+            collection = collection.filter(
+                Filter[f.type].apply(this, f.arguments)
+            )
+        })
+
+        if (mosaic) {
+            // Composite all images inn a collection (e.g. per country)
+            return collection.mosaic()
+        }
+
+        // There should only be one image after applying the filters
+        return Image(collection.first())
+    }
+
+    // Select band(s)
+    selectBand(eeImage) {
+        const { band, bandReducer } = this.options
+        const { Reducer } = this.ee
+
+        if (band) {
+            eeImage = eeImage.select(band)
+
+            if (Array.isArray(band) && bandReducer && Reducer[bandReducer]) {
+                // Combine multiple bands (e.g. age groups)
+                eeImage = eeImage.reduce(Reducer[bandReducer]())
+            }
         }
 
         return eeImage
     }
 
-    createLegend() {
-        const params = this.options.params
-        const min = params.min
-        const max = params.max
-        const palette = params.palette.split(',')
-        const step = (params.max - min) / (palette.length - (min > 0 ? 2 : 1))
-
-        let from = min
-        let to = Math.round(min + step)
-
-        return palette.map((color, index) => {
-            const item = { color }
-
-            if (index === 0 && min > 0) {
-                // Less than min
-                item.from = 0
-                item.to = min
-                item.name = '< ' + item.to
-                to = min
-            } else if (from < max) {
-                item.from = from
-                item.to = to
-                item.name = item.from + ' - ' + item.to
-            } else {
-                // Higher than max
-                item.from = from
-                item.name = '> ' + item.from
-            }
-
-            from = to
-            to = Math.round(min + step * (index + (min > 0 ? 1 : 2)))
-
-            return item
-        })
-    }
-
-    applyFilter(collection, filterOpt) {
-        const filter = filterOpt || this.options.filter
-
-        if (filter) {
-            filter.forEach(item => {
-                collection = collection.filter(
-                    this._ee.Filter[item.type].apply(this, item.arguments)
-                ) // eslint-disable-line
-            })
-        }
-
-        return collection
-    }
-
     // Run methods on image
-    runMethods(image) {
-        const methods = this.options.methods
-        let eeImage = image
+    runMethods(eeImage) {
+        const { methods } = this.options
 
         if (methods) {
             Object.keys(methods).forEach(method => {
                 if (eeImage[method]) {
-                    // Make sure method exist
-                    eeImage = eeImage[method].apply(eeImage, methods[method]) // eslint-disable-line
+                    eeImage = eeImage[method].apply(eeImage, methods[method])
                 }
             })
         }
@@ -211,13 +269,29 @@ class EarthEngine extends Layer {
         return eeImage
     }
 
+    // Mask out 0-values
+    maskImage(eeImage) {
+        return this.options.mask ? eeImage.updateMask(eeImage.gt(0)) : eeImage
+    }
+
     // Classify image according to legend
     classifyImage(eeImage) {
-        const legend = this._legend
+        const { legend = [], params } = this.options
         let zones
 
-        for (let i = 0, item; i < legend.length - 1; i++) {
+        if (!params) {
+            // Image has classes (e.g. landcover)
+            this.params = getParamsFromLegend(legend)
+            return eeImage
+        }
+
+        const min = 0
+        const max = legend.length - 1
+        const { palette } = params
+
+        for (let i = min, item; i < max; i++) {
             item = legend[i]
+
             if (!zones) {
                 zones = eeImage.gt(item.to)
             } else {
@@ -225,70 +299,53 @@ class EarthEngine extends Layer {
             }
         }
 
+        // Visualisation params
+        this.params = { min, max, palette }
+
         return zones
     }
 
     // Visualize image (turn into RGB)
     visualize(eeImage) {
-        const { legend, params } = this.options
+        const { params } = this.options
+
+        // Clip image to org unit features
+        if (this.featureCollection) {
+            eeImage = eeImage.clipToCollection(this.featureCollection)
+        }
 
         return new Promise(resolve =>
-            eeImage
-                .visualize(
-                    legend
-                        ? params
-                        : {
-                              min: 0,
-                              max: this._legend.length - 1,
-                              palette: params.palette,
-                          }
-                )
-                .getMap(null, resolve)
+            eeImage.visualize(this.params || params).getMap(null, resolve)
         )
     }
 
-    // Returns value at location in a callback
-    getValue(latlng, callback) {
-        const point = this._ee.Geometry.Point(latlng.lng, latlng.lat) // eslint-disable-line
-        const options = this.options
-        let dictionary
+    // Returns value at at position
+    getValue = latlng => {
+        const { band, legend } = this.options
+        const { Geometry, Reducer } = this.ee
+        const { lng, lat } = latlng
+        const point = Geometry.Point(lng, lat)
 
-        if (options.aggregation === 'mosaic') {
-            dictionary = this.eeImage.reduceRegion(
-                this._ee.Reducer.mean(),
-                point,
-                options.resolution,
-                options.projection
-            ) // eslint-disable-line
-        } else {
-            dictionary = this.eeImage.reduceRegion(
-                this._ee.Reducer.mean(),
-                point
-            ) // eslint-disable-line
-        }
+        return getInfo(
+            this.eeImage.reduceRegion(Reducer.mean(), point, 1)
+        ).then(data => {
+            const value = data[band] || Object.values(data)[0]
 
-        dictionary.getInfo(valueObj => {
-            const band = options.band || Object.keys(valueObj)[0]
-            let value = valueObj[band]
+            // Used for landcover
+            const item =
+                Array.isArray(legend) && legend.find(i => i.id === value)
 
-            if (options.legend && options.legend[value]) {
-                value = options.legend[value].name
-            } else if (options.value) {
-                // Needs calculation
-                value = options.value(value)
-            }
-
-            callback(value)
+            return item ? item.name : value
         })
     }
 
     // TODO: Move popup handling to the maps app
-    showValue = latlng => {
-        this.getValue(latlng, value => {
+    showValue = latlng =>
+        this.getValue(latlng).then(value => {
             const { lng, lat } = latlng
             const options = {
                 ...this.options,
-                value,
+                value: typeof value === 'number' ? setPrecision(value) : value,
             }
 
             const content = options.popup.replace(
@@ -298,6 +355,66 @@ class EarthEngine extends Layer {
 
             this._map.openPopup(document.createTextNode(content), [lng, lat])
         })
+
+    // Returns ee image when ready
+    getImage = () =>
+        new Promise(resolve =>
+            this.eeImage
+                ? resolve(this.eeImage)
+                : this.once('imageready', evt => resolve(evt.image))
+        )
+
+    // Perform aggregations to org unit features
+    aggregate = async aggregationType => {
+        const { legend } = this.options
+        const classes = hasClasses(aggregationType)
+        const image = await this.getImage()
+        const collection = this.featureCollection
+        const scale = this.scale
+        const { Reducer } = this.ee
+
+        if (classes && legend) {
+            // Used for landcover
+            const reducer = Reducer.frequencyHistogram()
+
+            return getInfo(
+                image
+                    .reduceRegions(collection, reducer, scale)
+                    .select(['histogram'], null, false)
+            ).then(data =>
+                getHistogramStatistics({
+                    data,
+                    scale,
+                    aggregationType,
+                    legend,
+                })
+            )
+        } else if (Array.isArray(aggregationType) && aggregationType.length) {
+            const reducer = combineReducers(ee)(aggregationType)
+
+            const aggFeatures = image
+                .reduceRegions({
+                    collection,
+                    reducer,
+                    scale,
+                })
+                .select(aggregationType, null, false)
+
+            return getInfo(aggFeatures).then(getFeatureCollectionProperties)
+        }
+    }
+
+    setOpacity(opacity) {
+        super.setOpacity(opacity)
+
+        const id = this.getId()
+        const layerId = `${id}-polygon`
+        const mapgl = this.getMapGL()
+
+        if (mapgl.getLayer(layerId)) {
+            // Clickable polygon layer should always be transparent
+            mapgl.setPaintProperty(layerId, 'fill-opacity', 0)
+        }
     }
 }